chore: moved database endpoints to main.py
