+++ main.py
@@ -7,19 +7,20 @@ from routing.load_data import load_flooded_areas
 from tsp_endpoint import main_tsp
 from tests.naive_tsp import naive_tsp  # For testing
 
-from database_endpoints import login_endpoint, convert_coordinates_endpoint
-
 from routing.route_directions import directions
 from models import DirectionsRequest
 from routing.cache_database import (
     connect_to_database,
     close_database_connection,
+    search_login,
     route_info,
     update_rescued_boolean,
 )
-from models import Point, RouteInfo, UpdateRescued
+from models import Point, LoginInput, RouteInfo, UpdateRescued
 from qc_coordinates import check_point_in_polygon
 from own_websocket import own_socket
+from db_env import GOOGLE_MAPS_API
+import googlemaps
 
 
 # Load the flooded areas on startup
@@ -34,13 +35,12 @@ async def startup_event(app: FastAPI) -> AsyncGenerator[None, None]:
 
 # Initialize the FastAPI app
 app = FastAPI(lifespan=startup_event)
+gmaps = googlemaps.Client(key = GOOGLE_MAPS_API)
 
 # Include router for tsp endpoint
 app.include_router(main_tsp.router)
 app.include_router(naive_tsp.router)  # For testing
 app.include_router(own_socket.router)
-app.include_router(login_endpoint.router)
-app.include_router(convert_coordinates_endpoint.router)
 
 
 # app.include_router(route_directions.router)
@@ -67,6 +67,30 @@ async def test():
         json_data = json.load(f)
     return json_data
 
+@app.post("/login", status_code = status.HTTP_200_OK)
+async def login(login_input: LoginInput):
+    db_data = await search_login(login_input.username, login_input.password)
+    
+    # Valid
+    if db_data: 
+        return db_data
+
+    # Invalid: Return person_id = 0, access_level = 0
+    # since these cases are normally impossible.
+    else:
+        return {
+            "person_id": 0,
+            "access_control": 0
+        }
+
+@app.post("/convert_coordinates", status_code = status.HTTP_200_OK)
+async def convert_coordinates(points: list[Point]):
+    location_names: list[str]= list()
+    for point in points:
+        print(point)
+        lng, lat = point.coordinates[0], point.coordinates[1]
+        location_names.append(gmaps.reverse_geocode((lat, lng), result_type="street_address")[0]["formatted_address"])
+    return {"locations": location_names}
 
 @app.post("/get_route_info", status_code=status.HTTP_200_OK)
 async def get_route_info(route_id: RouteInfo):
